Extract reading config to a separate directory
Shortens index.ts
