+++ src/cli/cli-dependencies.ts
@@ -1,9 +1,10 @@
 import { CliOptions } from './cli-options';
-import { saveJSONFile } from './io';
+import { readJSONArray, saveJSONFile } from './io';
 
 export interface CliDependencies {
   log: typeof console.log;
   saveJSONFile: typeof saveJSONFile;
+  readJSONArray: typeof readJSONArray;
   cliOptions: CliOptions;
 }
 
@@ -13,4 +14,5 @@ export const getCliDependencies = (
   cliOptions,
   log: console.log.bind(console),
   saveJSONFile,
+  readJSONArray,
 });
