+++ src/cli/config/index.ts
@@ -1,2 +1,3 @@
 export * from './initialize-config-files';
 export * from './update-loosely-type-checked-file-paths';
+export * from './read-config';
