+++ src/cli/io/index.ts
@@ -1,1 +1,3 @@
 export * from './save-json-file';
+export * from './read-json-array';
+export * from './get-program-input';
