+++ src/cli/io/read-json-array.ts
@@ -1,5 +1,4 @@
 import { readFileSync } from 'fs';
-import { createInterface } from 'readline';
 
 export const readJSONArray = (
   path: string,
@@ -30,17 +29,3 @@ export const readJSONArray = (
 
   return parsedFile;
 };
-
-export const getProgramInput = () =>
-  new Promise<string[]>((resolve) => {
-    const programInput: string[] = [];
-    const rl = createInterface(process.stdin);
-
-    rl.on('line', (line) => {
-      programInput.push(line);
-    });
-
-    rl.once('close', () => {
-      resolve(programInput);
-    });
-  });
